разделяет js-код на отдельные модули
